+++ src/HopDocusign.php
@@ -37,7 +37,7 @@ use yii\base\Event;
  *
  * @author    Hop Studios
  * @package   HopDocusign
- * @since     1.0.0
+ * @since     1.1.0
  *
  * @property  HopDocusignServiceService $hopDocusignService
  */
@@ -57,7 +57,7 @@ class HopDocusign extends Plugin
     /**
      * @var string
      */
-    public $schemaVersion = '1.0.0';
+    public $schemaVersion = '1.1.0';
 
     /**
      * @var bool
@@ -162,26 +162,6 @@ class HopDocusign extends Plugin
 
                 $fields = $form->getLayout()->getFields();
 
-                // // Iterate over all posted fields and get their values
-                // foreach ($fields as $field) {
-
-                //     // Bypass fields such as HTML or Submit, etc.
-                //     if ($field instanceof NoStorageInterface) {
-                //         continue;
-                //     }
-
-                //     // $field->getValue();
-                //     echo '<pre>';
-                //     print_r($field->getType());
-                //     echo '</pre>';
-                //     echo '<pre>';
-                //     print_r($field->getHandle());
-                //     echo '</pre>';
-                //     echo '<pre>';
-                //     var_dump($field->getValue());
-                //     echo '</pre>';
-                // }
-
                 require_once('Library/docusign/esign-client/autoload.php');
 
                 // Let the magic begin...
@@ -254,8 +234,10 @@ class HopDocusign extends Plugin
                                     'radioGroupTabs' => array()
                                 );
 
-                                // We're going to loop through the fields now. Some of them may be used to build the recipient name
-                                $template_role_name = $template->recipient_name;
+                                // We're going to loop through the fields now. Some of them may be used to build the template information
+                                $template_role_email = $template->template_role_email;
+                                $template_role_name = $template->template_role_name;
+                                $subject_line = $template->email_subject;
 
                                 // Assigning fields into their 'tab', we're only using 4 types of field here so 4 tabs
                                 foreach ($fields as $field) {
@@ -275,6 +257,9 @@ class HopDocusign extends Plugin
                                             $value = !empty($value[0]) ? $value[0] : trim(implode(' ', $value));
                                             $emailTab->setValue($value);
                                             $data['emailTabs'][] = $emailTab;
+
+                                            // Replace if key is used to build the template role email
+                                            $template_role_email = str_replace('{' . $key . '}', $value, $template_role_email);
                                             break;
                                         case 'text':
                                             $textTab = new \DocuSign\eSign\Model\Text();
@@ -284,6 +269,7 @@ class HopDocusign extends Plugin
 
                                             // Replace if key is used to build the recipient name
                                             $template_role_name = str_replace('{' . $key . '}', $value, $template_role_name);
+                                            $subject_line = str_replace('{' . $key . '}', $value, $subject_line);
                                             break;
                                         case 'checkbox':
                                             $checkboxTab = new \DocuSign\eSign\Model\Checkbox();
@@ -312,9 +298,7 @@ class HopDocusign extends Plugin
                                 $tabs->setEmailTabs($data['emailTabs']);
                                 $tabs->setTextTabs($data['textTabs']);
 
-                                // Setting the email and name for the template role
-                                $template_role_email = implode(',', $form->get($template->email_handle)->getValue());
-
+                                // Setting template role information
                                 $templateRole->setEmail($template_role_email);
                                 $templateRole->setName($template_role_name);
 
@@ -323,7 +307,14 @@ class HopDocusign extends Plugin
 
                                 // Instantiate a new envelope object and configure settings
                                 $envelop_definition = new \DocuSign\eSign\Model\EnvelopeDefinition();
-                                $envelop_definition->setEmailSubject($template->email_subject . $clientUserId);
+
+                                if (strpos($subject_line, '{clientUserId}')) {
+                                    $subject_line = str_replace('{clientUserId}', $clientUserId, $subject_line);
+                                }
+
+                                if (!empty($subject_line)) {
+                                    $envelop_definition->setEmailSubject($subject_line);
+                                }
 
                                 // Set which template to create and which role is signing
                                 $envelop_definition->setTemplateId($template->template_id);