+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Hop DocuSign Changelog
 
+## 1.1.0 - 2020-06-22
+### Updated
+- {clientUserId} for email title
+- Template Role Email dynamic generation
+- Subject line dynamic generation
+- Code clean-up
+
+## 1.0.3 - 2020-06-19
+### Removed
+- Unecessary code
+
 ## 1.0.1 - 2020-06-15
 ### Fixed
 - Only modify the signing url when returned